CageConfiguration to utils file
